Automate actually printing the crossword
Because macOS is POSIX-compatible, we can use `lp` to print from the
terminal. If passed a file with no other arguments, it will print to our
default printer, which is exactly what we want.

I'm not *wild* about the `if curl` line, but not bothered enough to
actually fix it right now .In future, it might be nice if it emailed
failure notifications to me.